Chore: 함수 분리 및 제거
