+++ src/utils/getTimeDifference.ts
@@ -1,28 +1,25 @@
-const getTimeDifference = (createdAt: string) => {
-  const date = new Date(createdAt);
-  const currentDate = new Date();
-  const timeDifference = Math.floor((currentDate.getTime() - date.getTime()) / 1000);
-  const minute = Math.floor(timeDifference / 60);
-  if (!minute || minute < 0) {
-    return 'InvalidTime';
+const getTimeDifference = (startsAt: string, workhour: number) => {
+  const startDate = new Date(startsAt);
+  const localStartDate = new Date(startDate.getTime() - 9 * 60 * 60000);
+  const localEndDate = new Date(localStartDate.getTime() + 60000 * workhour * 60);
+
+  const starYear = localStartDate.getFullYear();
+  const startMonth = String(localStartDate.getMonth() + 1).padStart(2, '0');
+  const startDay = String(localStartDate.getDate()).padStart(2, '0');
+  const startHour = String(localStartDate.getHours()).padStart(2, '0');
+  const startMinute = String(localStartDate.getMinutes()).padStart(2, '0');
+
+  const endYear = localEndDate.getFullYear();
+  const endMonth = String(localEndDate.getMonth() + 1).padStart(2, '0');
+  const endDay = String(localEndDate.getDate()).padStart(2, '0');
+  const endHour = String(localEndDate.getHours()).padStart(2, '0');
+  const endMinute = String(localEndDate.getMinutes()).padStart(2, '0');
+  const endFormattedDate = `${endYear}.${endMonth}.${endDay} ${endHour}:${endMinute}`;
+  const formattedDate = `${starYear}.${startMonth}.${startDay} ${startHour}:${startMinute}`;
+  if (endDay !== startDay) {
+    return `${formattedDate}~${endFormattedDate}`;
   }
-  if (minute < 60) {
-    return `${minute}분 전`;
-  }
-  if (minute < 60 * 24) {
-    const hour = Math.floor(minute / 60);
-    return `${hour}시간 전`;
-  }
-  if (minute < 60 * 24 * 30) {
-    const day = Math.floor(minute / (60 * 24));
-    return `${day}일 전`;
-  }
-  if (minute < 60 * 24 * 30 * 12) {
-    const month = Math.floor(minute / (60 * 24 * 30));
-    return `${month}달 전`;
-  }
-  const year = Math.floor(minute / (60 * 24 * 30 * 12));
-  return `${year}년 전`;
+  return `${formattedDate}~${endHour}:${endMinute}`;
 };
 
 export default getTimeDifference;
